Add files via upload
Takes a 2D array of 3-digit integers and an empty 1D integer array that has the same capacity as the row count of the 2D array. The function sums the middle digits of all elements in a given row, and the sum for each row is stored in the index of the 1D array corresponding to that row index in the 2D array.